VC warning fixes
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk/ac@4718 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ source/src/oggstream.cpp
@@ -9,7 +9,7 @@
 static size_t oggcallbackread(void *ptr, size_t size, size_t nmemb, void *datasource)
 {
     stream *s = (stream *)datasource;
-    return s ? s->read(ptr, size*nmemb)/size : 0;
+    return s ? s->read(ptr, int(size*nmemb))/size : 0;
 }
 
 static int oggcallbackseek(void *datasource, ogg_int64_t offset, int whence)
